+++ set-env.redshift.template.sh
@@ -1,6 +1,3 @@
 export STORAGE_API_URL=https://connection-testing.keboola.com/
 export STORAGE_API_TOKEN=
 export STORAGE_API_MAINTENANCE_URL=https://maintenance-testing.keboola.com/
-export REDSHIFT_HOSTNAME=
-export REDSHIFT_USER=
-export REDSHIFT_PASSWORD=