+++ src/Keboola/StorageApi/Workspaces.php
@@ -9,13 +9,6 @@
 
 namespace Keboola\StorageApi;
 
-use Keboola\StorageApi\Options\Components\Configuration;
-use Keboola\StorageApi\Options\Components\ConfigurationRow;
-use Keboola\StorageApi\Options\Components\ListConfigurationRowsOptions;
-use Keboola\StorageApi\Options\Components\ListConfigurationRowVersionsOptions;
-use Keboola\StorageApi\Options\Components\ListConfigurationsOptions;
-use Keboola\StorageApi\Options\Components\ListConfigurationVersionsOptions;
-
 class Workspaces
 {