+++ tests/Backend/CommonPart2/SlicedImportsTest.php
@@ -7,7 +7,7 @@
  *
  */
 
-namespace Keboola\Test\Backend\Common;
+namespace Keboola\Test\Backend\CommonPart2;
 
 use Keboola\Test\StorageApiTestCase;