+++ tests/Backend/CommonPart2/TableExporterTest.php
@@ -8,7 +8,7 @@
  */
 
 
-namespace Keboola\Test\Backend\Common;
+namespace Keboola\Test\Backend\CommonPart2;
 
 use Keboola\Test\StorageApiTestCase;
 use Keboola\StorageApi\Client;